requirements.txt
streamlit
rembg
Pillow